class/future_content/hackathon/script/requirements.txt
torch
torchvision
pillow
requests
